requirements.txt
sqlalchemy
fastapi
pydantic_settings
uvicorn[standard]
aiosqlite
asyncpg
pydantic_settings
pytest
python-jose[cryptography] 
passlib[bcrypt]
pyotp